refactor: extract submission table component
