+++ Web Ui/src/sections/Assignments/AssignmentDetail.tsx
@@ -7,11 +7,6 @@ import {
   Card,
   CardContent,
   Divider,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
   Typography,
 } from "@mui/material";
 import CalendarMonthIcon from "@mui/icons-material/CalendarMonth";
@@ -20,17 +15,18 @@ import GroupsIcon from "@mui/icons-material/Groups";
 import {
   AccessTime as AccessTimeIcon,
   Link as LinkIcon,
-  RemoveCircle as RemoveCircleIcon,
   Comment as CommentIcon,
 } from "@mui/icons-material";
 import { GitLinkDialog } from "./components/GitHubLinkDialog";
 import { CommentDialog } from "./components/CommentDialog";
 import CircularProgress from "@mui/material/CircularProgress";
 import { InfoRow } from "../Shared/Components/InfoRow";
+import { SubmissionTable } from "./components/SubmissionTable";
 import SubmissionRepository from "../../modules/Submissions/Repository/SubmissionRepository";
 import { CreateSubmission } from "../../modules/Submissions/Aplication/createSubmission";
 import {
   SubmissionCreationObject,
+  SubmissionDataObject,
   SubmissionUpdateObject,
 } from "../../modules/Submissions/Domain/submissionInterfaces";
 import UsersRepository from "../../modules/Users/repository/UsersRepository";
@@ -241,155 +237,31 @@ const AssignmentDetail: React.FC<AssignmentDetailProps> = ({
         return status;
     }
   };
-
-  const submissionRows = submissions.map((submission) => {
-    const studentEmail = studentEmails[submission.userid] ?? "";
-    const formattedStartDate = formatDate(submission.start_date.toString());
-    const formattedEndDate = submission.end_date
-      ? formatDate(submission.end_date.toString())
-      : "N/A";
-    const hasRepositoryLink = submission.repository_link !== "";
-    const teacherStatus = hasRepositoryLink ? "Enviado" : "No enviado";
-    const statusColor = hasRepositoryLink ? "#4CAF50" : "#F44336";
-
-    return (
-      <TableRow key={submission.id} sx={{ borderBottom: "1px solid #C9C9C9" }}>
-        <TableCell
-          sx={{
-            py: 2.2,
-            fontSize: "1.3rem",
-            maxWidth: 0,
-            overflow: "hidden",
-            textOverflow: "ellipsis",
-            whiteSpace: "nowrap",
-          }}
-          title={studentEmail}
-        >
-          {studentEmail}
-        </TableCell>
-        <TableCell
-          sx={{
-            py: 2.2,
-            fontSize: "1.3rem",
-            color: statusColor,
-            borderLeft: "1px solid #C9C9C9",
-            whiteSpace: "nowrap",
-          }}
-        >
-          {teacherStatus}
-        </TableCell>
-        <TableCell
-          sx={{ py: 2.2, borderLeft: "1px solid #C9C9C9", textAlign: "center" }}
-        >
-          {hasRepositoryLink ? (
-            <a
-              href={submission.repository_link}
-              target="_blank"
-              rel="noopener noreferrer"
-              style={{
-                display: "inline-flex",
-                alignItems: "center",
-                color: "#5C6BC0",
-              }}
-            >
-              <LinkIcon />
-            </a>
-          ) : (
-            <RemoveCircleIcon sx={{ color: "#F44336" }} />
-          )}
-        </TableCell>
-        <TableCell
-          sx={{ py: 2.2, fontSize: "1.3rem", borderLeft: "1px solid #C9C9C9" }}
-        >
-          {formattedStartDate}
-        </TableCell>
-        <TableCell
-          sx={{ py: 2.2, fontSize: "1.3rem", borderLeft: "1px solid #C9C9C9" }}
-        >
-          {formattedEndDate}
-        </TableCell>
-        <TableCell sx={{ py: 2.2, borderLeft: "1px solid #C9C9C9" }}>
-          <Button
-            variant="contained"
-            disabled={submission.repository_link === ""}
-            onClick={() => {
-              localStorage.setItem("selectedMetric", "Dashboard");
-              handleRedirectAdmin(
-                submission.repository_link,
-                submissions,
-                submission.id,
-                "/graph"
-              );
-            }}
-            color="primary"
-            style={{
-              textTransform: "none",
-              fontSize: "1.15rem",
-              marginRight: "8px",
-              backgroundColor:
-                submission.repository_link === "" ? "#BDBDBD" : undefined,
-              minWidth: "110px",
-            }}
-          >
-            Ver grafica
-          </Button>
-        </TableCell>
-
-        <TableCell sx={{ py: 2.2, borderLeft: "1px solid #C9C9C9" }}>
-          <Button
-            variant="contained"
-            disabled={submission.repository_link === ""}
-            onClick={() => {
-              navigate("/asistente-ia", {
-                state: { repositoryLink: submission.repository_link },
-              });
-            }}
-            color="primary"
-            style={{
-              textTransform: "none",
-              fontSize: "1.15rem",
-              marginRight: "8px",
-              backgroundColor:
-                submission.repository_link === "" ? "#BDBDBD" : undefined,
-              minWidth: "110px",
-            }}
-          >
-            Asistente
-          </Button>
-        </TableCell>
-        {!isStudent(role) && (
-          <TableCell sx={{ py: 2.2, borderLeft: "1px solid #C9C9C9" }}>
-            <Button
-              variant="contained"
-              disabled={submission.repository_link === "" || disableAdditionalGraphs}
-              onClick={() => {
-                localStorage.setItem("selectedMetric", "Complejidad");
-                handleRedirectAdmin(
-                  submission.repository_link,
-                  submissions,
-                  submission.id,
-                  "/aditionalgraph"
-                );
-              }}
-              color="primary"
-              style={{
-                textTransform: "none",
-                fontSize: "1.15rem",
-                marginRight: "7px",
-                backgroundColor:
-                  submission.repository_link === "" || disableAdditionalGraphs
-                    ? "#BDBDBD"
-                    : undefined,
-                minWidth: "84px",
-              }}
-            >
-              Ver
-            </Button>
-          </TableCell>
-        )}
-      </TableRow>
+  const handleViewGraph = (targetSubmission: SubmissionDataObject) => {
+    localStorage.setItem("selectedMetric", "Dashboard");
+    handleRedirectAdmin(
+      targetSubmission.repository_link,
+      submissions,
+      targetSubmission.id,
+      "/graph"
     );
-  });
+  };
+
+  const handleOpenAssistant = (targetSubmission: SubmissionDataObject) => {
+    navigate("/asistente-ia", {
+      state: { repositoryLink: targetSubmission.repository_link },
+    });
+  };
+
+  const handleViewAdditionalGraph = (targetSubmission: SubmissionDataObject) => {
+    localStorage.setItem("selectedMetric", "Complejidad");
+    handleRedirectAdmin(
+      targetSubmission.repository_link,
+      submissions,
+      targetSubmission.id,
+      "/aditionalgraph"
+    );
+  };
 
 
   return (
@@ -611,23 +483,15 @@ const AssignmentDetail: React.FC<AssignmentDetailProps> = ({
                 <CircularProgress size={40} thickness={4} />
               </div>
             ) : (
-              <Table sx={{ borderCollapse: "collapse", tableLayout: "fixed" }}>
-                <TableHead>
-                  <TableRow sx={{ borderBottom: "1px solid #C9C9C9" }}>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "18%" }}>Correo</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "11%", borderLeft: "1px solid #C9C9C9" }}>Estado</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "10%", borderLeft: "1px solid #C9C9C9" }}>Enlace</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "13%", borderLeft: "1px solid #C9C9C9" }}>Fecha de Inicio</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "14%", borderLeft: "1px solid #C9C9C9" }}>Fecha de Finalización</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "13%", borderLeft: "1px solid #C9C9C9" }}>Grafica</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "12%", borderLeft: "1px solid #C9C9C9" }}>Asistente AI</TableCell>
-                    <TableCell sx={{ fontWeight: 700, fontSize: "1.65rem", py: 1.3, width: "12%", borderLeft: "1px solid #C9C9C9" }}>Graficas Adicionales</TableCell>
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  {submissionRows}
-                </TableBody>
-              </Table>
+              <SubmissionTable
+                submissions={submissions}
+                studentEmails={studentEmails}
+                disableAdditionalGraphs={disableAdditionalGraphs}
+                showAdditionalGraphs={!isStudent(role)}
+                onViewGraph={handleViewGraph}
+                onOpenAssistant={handleOpenAssistant}
+                onViewAdditionalGraph={handleViewAdditionalGraph}
+              />
             )}
         </div>
       )}
